Rename entry file to be JSX
