+++ ui/lerna.json
@@ -3,12 +3,13 @@
   "npmClient": "yarn",
   "tagVersionPrefix": "ui-v",
   "useWorkspaces": true,
-  "packages": [
-    "packages/shared/*"
-  ],
+  "packages": ["packages/shared/*"],
   "command": {
     "publish": {
       "conventionalCommits": true
+    },
+    "version": {
+      "allowBranch": "main"
     }
   }
 }
