+++ ui/packages/shared/store/src/hooks.ts
@@ -1,5 +1,5 @@
 import {TypedUseSelectorHook, useDispatch, useSelector} from 'react-redux';
-import type {RootState, AppDispatch} from '.';
+import type {RootState, AppDispatch} from './store';
 
 // Use throughout app instead of plain `useDispatch` and `useSelector`
 export const useAppDispatch = () => useDispatch<AppDispatch>();
