+++ ui/packages/shared/store/src/index.tsx
@@ -1,4 +1,4 @@
-import store from './index';
+import store from './store';
 import {useAppDispatch, useAppSelector} from './hooks';
 import {selectDarkMode, setDarkMode} from './slices/uiSlice';
 import {selectCompareMode, setCompare} from './slices/profileSlice';
